refactor: layout 파일명을 케밥 케이스로 변경 (#447)
