+++ data/project/timegambit-blog.mdx
@@ -2,7 +2,7 @@
 title: Time Gambit 블로그
 summary: Time Gambit이 만든 블로그를 소개합니다.
 images: [https://i.imgur.com/VIFO53Q.png]
-layout: ProjectLayout
+layout: PostLayoutWithThumbnail
 ---
 
 ## 프로젝트 소개
