+++ src/app/series/[slug]/page.tsx
@@ -11,7 +11,7 @@ import { pickBlogItem, sortedBlogPost } from '@/lib/contentlayer';
 import { getAllSeries } from '@/lib/get-blog-info.mjs';
 import { getImageWithFallback } from '@/lib/utils';
 
-import ListLayout from '@/layouts/ListLayout';
+import ListLayout from '@/layouts/list-layout';
 import { SeriesListItem } from '@/types';
 
 interface Props {
